Attended PR comment - rename packages
