+++ apps/connect/source/ftrack_connect/application_launcher/__init__.py
