+++ apps/connect/source/ftrack_connect/plugin_manager/overlay.py
