+++ apps/connect/source/ftrack_connect/plugin_manager/processor.py
